locate some json conversion code with the types, clean up unit tests after removal of helpers
