+++ Core/include/Acts/Utilities/BinUtility.hpp
@@ -328,4 +328,8 @@ class BinUtility {
   Transform3 m_itransform;                 /// unique inverse transform
 };
 
+void to_json(nlohmann::json& j, const BinUtility& bu);
+
+void from_json(const nlohmann::json& j, BinUtility& bu);
+
 }  // namespace Acts
