+++ Core/src/Definitions/CMakeLists.txt
@@ -1,1 +1,4 @@
-target_sources(ActsCore PRIVATE Common.cpp Direction.cpp ParticleData.cpp)
+target_sources(
+    ActsCore
+    PRIVATE Common.cpp Direction.cpp ParticleData.cpp Algebra.cpp
+)
