+++ Core/src/Geometry/Extent.cpp
@@ -211,3 +211,51 @@ std::ostream& Acts::operator<<(std::ostream& sl, const Extent& rhs) {
   sl << rhs.toString();
   return sl;
 }
+
+void Acts::to_json(nlohmann::json& j, const Acts::Extent& e) {
+  {
+    nlohmann::json jrange;
+    const auto& xrange = e.range();
+    for (auto ibv : allBinningValues()) {
+      if (e.constrains(ibv)) {
+        jrange[binningValueName(ibv)] = xrange[toUnderlying(ibv)];
+      }
+    }
+    j["range"] = jrange;
+  }
+
+  {
+    nlohmann::json jenvelope;
+    const auto& envelope = e.envelope();
+    for (auto ibv : allBinningValues()) {
+      if (envelope[ibv] != zeroEnvelope) {
+        jenvelope[binningValueName(ibv)] =
+            Range1D<double>(envelope[ibv][0], envelope[ibv][1]);
+      }
+    }
+    if (!jenvelope.empty()) {
+      j["envelope"] = jenvelope;
+    }
+  }
+}
+
+void Acts::from_json(const nlohmann::json& j, Acts::Extent& e) {
+  const auto& jrange = j["range"];
+
+  for (const auto& [key, value] : jrange.items()) {
+    BinningValue bval = binningValueFromName(key);
+    e.set(bval, value["min"], value["max"]);
+  }
+
+  if (j.contains("envelope")) {
+    const auto& jenvelope = j["envelope"];
+    ExtentEnvelope envelope;
+
+    for (const auto& [key, value] : jenvelope.items()) {
+      BinningValue bval = binningValueFromName(key);
+      envelope[bval] = {value["min"], value["max"]};
+    }
+
+    e.setEnvelope(envelope);
+  }
+}
