+++ Plugins/Json/include/Acts/Plugins/Json/AlgebraJsonConverter.hpp
@@ -17,10 +17,6 @@
 // can not match our naming guidelines.
 namespace Acts {
 
-void to_json(nlohmann::json& j, const Transform3& t);
-
-void from_json(const nlohmann::json& j, Transform3& t);
-
 namespace Transform3JsonConverter {
 
 /// @brief The options for the transform converter
