+++ Plugins/Json/src/AlgebraJsonConverter.cpp
@@ -11,43 +11,6 @@
 #include <array>
 #include <stdexcept>
 
-void Acts::to_json(nlohmann::json& j, const Acts::Transform3& t) {
-  auto translation = t.translation();
-  if (translation != Acts::Vector3(0., 0., 0)) {
-    std::array<double, 3> tdata = {translation.x(), translation.y(),
-                                   translation.z()};
-    j["translation"] = tdata;
-  } else {
-    j["translation"] = nlohmann::json();
-  }
-
-  auto rotation = t.rotation();
-  if (rotation != Acts::RotationMatrix3::Identity()) {
-    std::array<double, 9> rdata = {
-        rotation(0, 0), rotation(0, 1), rotation(0, 2),
-        rotation(1, 0), rotation(1, 1), rotation(1, 2),
-        rotation(2, 0), rotation(2, 1), rotation(2, 2)};
-    j["rotation"] = rdata;
-  } else {
-    j["rotation"] = nlohmann::json();
-  }
-}
-
-void Acts::from_json(const nlohmann::json& j, Acts::Transform3& t) {
-  t = Acts::Transform3::Identity();
-  if (j.find("rotation") != j.end() && !j["rotation"].empty()) {
-    std::array<double, 9> rdata = j["rotation"];
-    Acts::RotationMatrix3 rot;
-    rot << rdata[0], rdata[1], rdata[2], rdata[3], rdata[4], rdata[5], rdata[6],
-        rdata[7], rdata[8];
-    t.prerotate(rot);
-  }
-  if (j.find("translation") != j.end() && !j["translation"].empty()) {
-    std::array<double, 3> tdata = j["translation"];
-    t.pretranslate(Acts::Vector3(tdata[0], tdata[1], tdata[2]));
-  }
-}
-
 nlohmann::json Acts::Transform3JsonConverter::toJson(const Transform3& t,
                                                      const Options& options) {
   nlohmann::json jTransform;
