+++ Plugins/Json/src/IndexedSurfacesJsonConverter.cpp
@@ -11,7 +11,6 @@
 #include "Acts/Navigation/NavigationStateUpdaters.hpp"
 #include "Acts/Plugins/Json/GridJsonConverter.hpp"
 #include "Acts/Plugins/Json/IndexedGridJsonHelper.hpp"
-#include "Acts/Plugins/Json/UtilitiesJsonConverter.hpp"
 #include "Acts/Utilities/GridAxisGenerators.hpp"
 
 #include <array>
