+++ Plugins/Json/src/ProtoDetectorJsonConverter.cpp
@@ -10,7 +10,6 @@
 
 #include "Acts/Detector/ProtoDetector.hpp"
 #include "Acts/Geometry/Extent.hpp"
-#include "Acts/Plugins/Json/ExtentJsonConverter.hpp"
 #include "Acts/Plugins/Json/SurfaceJsonConverter.hpp"
 #include "Acts/Surfaces/Surface.hpp"
 #include "Acts/Utilities/BinningData.hpp"
