+++ Tests/UnitTests/Core/Geometry/ExtentTests.cpp
@@ -186,6 +186,28 @@ BOOST_AUTO_TEST_CASE(DesignatedInitializers) {
   BOOST_CHECK(exp == act);
 }
 
+BOOST_AUTO_TEST_CASE(JsonExtentRoundtripTests) {
+  Extent e;
+  e.set(BinningValue::binR, 0, 200);
+  e.set(BinningValue::binZ, -50, 50);
+
+  nlohmann::json j;
+  j["extent"] = e;
+
+  std::cout << j.dump(2) << std::endl;
+
+  Extent eIn = j["extent"];
+
+  CHECK_CLOSE_ABS(eIn.min(BinningValue::binR), e.min(BinningValue::binR),
+                  10e-5);
+  CHECK_CLOSE_ABS(eIn.max(BinningValue::binR), e.max(BinningValue::binR),
+                  10e-5);
+  CHECK_CLOSE_ABS(eIn.min(BinningValue::binZ), e.min(BinningValue::binZ),
+                  10e-5);
+  CHECK_CLOSE_ABS(eIn.max(BinningValue::binZ), e.max(BinningValue::binZ),
+                  10e-5);
+}
+
 BOOST_AUTO_TEST_SUITE_END()
 
 }  // namespace Acts::Test
