+++ Tests/UnitTests/Plugins/Json/CMakeLists.txt
@@ -4,14 +4,12 @@ add_unittest(AlgebraJsonConverter AlgebraJsonConverterTests.cpp)
 add_unittest(DetectorJsonConverter DetectorJsonConverterTests.cpp)
 add_unittest(DetectorVolumeJsonConverter DetectorVolumeJsonConverterTests.cpp)
 add_unittest(DetectorVolumeFinderJsonConverter DetectorVolumeFinderJsonConverterTests.cpp)
-add_unittest(ExtentJsonConverter ExtentJsonConverterTests.cpp)
 add_unittest(GeometryHierarchyMapJsonConverter GeometryHierarchyMapJsonConverterTests.cpp)
 add_unittest(GridJsonConverter GridJsonConverterTests.cpp)
 add_unittest(MaterialJsonConverter MaterialJsonConverterTests.cpp)
 add_unittest(MaterialMapJsonConverter MaterialMapJsonConverterTests.cpp)
 add_unittest(PortalJsonConverter PortalJsonConverterTests.cpp)
 add_unittest(ProtoDetectorJsonConverter ProtoDetectorJsonConverterTests.cpp)
-add_unittest(UtilitiesJsonConverter UtilitiesJsonConverterTests.cpp)
 add_unittest(SurfaceBoundsJsonConverter SurfaceBoundsJsonConverterTests.cpp)
 add_unittest(SurfaceJsonConverter SurfaceJsonConverterTests.cpp)
 add_unittest(VolumeBoundsJsonConverter VolumeBoundsJsonConverterTests.cpp)
